refactor: move order-receipt PDF generation server-side (#27, #42)
The ~290-line pdf-lib receipt generator was (a) shipped to and executed
in the client browser and (b) copy-pasted byte-for-byte across both
order-detail pages.

- Extracted the generation into lib/generate-receipt.js (single source,
  pure pdf-lib, no DOM).
- Added GET /api/orders/[orderId]/receipt - auth-checked and
  ownership-checked (404 not 403 for orders you don't own, so it can't be
  used to probe IDs), returns application/pdf.
- Both pages now ship a ~20-line fetch-and-download helper instead of the
  full generator. pdf-lib no longer enters the client bundle for these
  routes. Each page dropped from 456 to 189 lines.
